Project envvar support, process management, other stuff

+++ bin/tag.sh
@@ -11,7 +11,7 @@ git fetch --tags
 
 TGT=${1-none}
 if [[ $TGT == "none" ]]; then
-  TGT=$(git describe --tags | sed -e 's/v//g')
+  TGT=$(git describe --match "v[0-9]*" --tags --abbrev=0 | sed -e 's/v//g')
   TGT=$(echo ${TGT} | awk -F. -v OFS=. '{$NF++;print}')
 fi
 if [[ ${TGT:0:1} == "v" ]]; then
@@ -20,19 +20,27 @@ fi
 
 echo $TGT
 
-sed -i.bak -e "s/version = \\\"[v]*[0-9]*[0-9]\.[0-9]*[0-9]\.[0-9]*[0-9]\\\"/version = \"${TGT}\"/g" ./main.go
-rm -f "./main.go.bak"
-sed -i.bak -e "s/Version: \\\"[v]*[0-9]*[0-9]\.[0-9]*[0-9]\.[0-9]*[0-9]\\\"/Version: \"${TGT}\"/g" ./app/cmd/lib.go
-rm -f ./app/cmd/lib.go.bak
-sed -i.bak -e "s/\\\"version\\\": \\\"[v]*[0-9]*[0-9]\.[0-9]*[0-9]\.[0-9]*[0-9]\\\"/\"version\": \"${TGT}\"/g" ./.projectforge/project.json
-rm -f "./.projectforge/project.json.bak"
+pat="^[0-9]"
+if [[ $TGT =~ $pat ]]; then
+  sed -i.bak -e "s/version = \\\"[v]*[0-9]*[0-9]\.[0-9]*[0-9]\.[0-9]*[0-9]\\\"/version = \"${TGT}\"/g" ./main.go
+  rm -f "./main.go.bak"
+  sed -i.bak -e "s/Version: \\\"[v]*[0-9]*[0-9]\.[0-9]*[0-9]\.[0-9]*[0-9]\\\"/Version: \"${TGT}\"/g" ./app/cmd/lib.go
+  rm -f ./app/cmd/lib.go.bak
+  sed -i.bak -e "s/\\\"version\\\": \\\"[v]*[0-9]*[0-9]\.[0-9]*[0-9]\.[0-9]*[0-9]\\\"/\"version\": \"${TGT}\"/g" ./.projectforge/project.json
+  rm -f "./.projectforge/project.json.bak"
+fi
 
 make build
 
 git add .
-git commit -m "v${TGT}" || true
 
-git tag "v${TGT}"
+if [[ $TGT =~ $pat ]]; then
+  git commit -m "v${TGT}" || true
+  git tag "v${TGT}"
+else
+  git commit -m "${TGT}" || true
+  git tag "${TGT}"
+fi
 
 git push
 git push --tags
